Compare ch directly
Accidentally, polymorphic (<) was used where it's not intended. It was
doing the right thing in this particular case, but in general it might
produce subtly wrong results, because it compares all fields of a
record. In this use case only `ch` should be compared though. It might
go wring in cases of comparing e.g.
`{ch='c'; d=R} > {ch='c'; d=L}`
which yields true, but these two records should be considered equal.
In this particular use case that's not an issue though, because it's
guaranteed that all elements of compared array will have different `ch`
fields.